requirements.txt
et-xmlfile==1.1.0
lxml==5.1.0
openpyxl==3.1.2
pandas==2.2.0
PuLP==2.8.0
pyarrow==15.0.0
pyodbc==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
